backend/requirements/base.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
tzdata
python-dotenv>=1.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
